+++ src/canopy/knowledge_base/record_encoder/openai.py
@@ -1,15 +1,8 @@
 from typing import List
-from tenacity import (
-    retry,
-    wait_random_exponential,
-    stop_after_attempt,
-    retry_if_exception_type,
-)
 from pinecone_text.dense.openai_encoder import OpenAIEncoder
 from canopy.knowledge_base.models import KBDocChunk, KBEncodedDocChunk, KBQuery
 from canopy.knowledge_base.record_encoder.dense import DenseRecordEncoder
 from canopy.models.data_models import Query
-from canopy.utils.openai_exceptions import OPEN_AI_TRANSIENT_EXCEPTIONS
 
 
 class OpenAIRecordEncoder(DenseRecordEncoder):
@@ -34,15 +27,9 @@ class OpenAIRecordEncoder(DenseRecordEncoder):
                         Defaults to 400.
             **kwargs: Additional arguments to pass to the underlying `pinecone-text. OpenAIEncoder`.
         """  # noqa: E501
-        encoder = OpenAIEncoder(model_name)
-        super().__init__(dense_encoder=encoder, batch_size=batch_size, **kwargs)
+        encoder = OpenAIEncoder(model_name, **kwargs)
+        super().__init__(dense_encoder=encoder, batch_size=batch_size)
 
-    @retry(
-        reraise=True,
-        wait=wait_random_exponential(min=1, max=10),
-        stop=stop_after_attempt(3),
-        retry=retry_if_exception_type(OPEN_AI_TRANSIENT_EXCEPTIONS),
-    )
     def encode_documents(self, documents: List[KBDocChunk]) -> List[KBEncodedDocChunk]:
         """
         Encode a list of documents, takes a list of KBDocChunk and returns a list of KBEncodedDocChunk.