+++ src/canopy/llm/base.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Union, Iterable, Optional, List
 
-from canopy.llm.models import Function, ModelParams
+from canopy.llm.models import Function
 from canopy.models.api_models import ChatResponse, StreamingChatChunk
 from canopy.models.data_models import Messages, Query
 from canopy.utils.config import ConfigurableMixin
@@ -9,13 +9,8 @@ from canopy.utils.config import ConfigurableMixin
 
 class BaseLLM(ABC, ConfigurableMixin):
     def __init__(self,
-                 model_name: str,
-                 *,
-                 model_params: Optional[ModelParams] = None,
-                 ):
+                 model_name: str):
         self.model_name = model_name
-        # TODO: consider removing altogether
-        self.default_model_params = model_params or ModelParams()
 
     @abstractmethod
     def chat_completion(self,
@@ -23,7 +18,7 @@ class BaseLLM(ABC, ConfigurableMixin):
                         *,
                         stream: bool = False,
                         max_tokens: Optional[int] = None,
-                        model_params: Optional[ModelParams] = None,
+                        model_params: Optional[dict] = None,
                         ) -> Union[ChatResponse, Iterable[StreamingChatChunk]]:
         pass
 
@@ -33,7 +28,7 @@ class BaseLLM(ABC, ConfigurableMixin):
                                function: Function,
                                *,
                                max_tokens: Optional[int] = None,
-                               model_params: Optional[ModelParams] = None
+                               model_params: Optional[dict] = None
                                ) -> dict:
         pass
 
@@ -43,7 +38,7 @@ class BaseLLM(ABC, ConfigurableMixin):
                                *,
                                stream: bool = False,
                                max_generated_tokens: Optional[int] = None,
-                               model_params: Optional[ModelParams] = None,
+                               model_params: Optional[dict] = None,
                                ) -> Union[ChatResponse,
                                           Iterable[StreamingChatChunk]]:
         pass
@@ -53,6 +48,6 @@ class BaseLLM(ABC, ConfigurableMixin):
                                 messages: Messages,
                                 *,
                                 max_generated_tokens: Optional[int] = None,
-                                model_params: Optional[ModelParams] = None,
+                                model_params: Optional[dict] = None,
                                 ) -> List[Query]:
         pass